+++ lib/core/networking/api_service.dart
@@ -2,8 +2,10 @@
 import 'package:dio/dio.dart';
 import 'package:doc_app/core/networking/api_constants.dart';
 
-import 'package:doc_app/feathures/login/logic/data/login_request_body.dart';
-import 'package:doc_app/feathures/login/logic/data/login_response.dart';
+import 'package:doc_app/feathures/login/data/login_request_body.dart';
+import 'package:doc_app/feathures/login/data/login_response.dart';
+import 'package:doc_app/feathures/sign_up/data/models/signup_response.dart';
+import 'package:doc_app/feathures/sign_up/data/models/sinup_request_body.dart';
 
 import 'package:retrofit/http.dart';
 import 'package:retrofit/retrofit.dart';
@@ -17,5 +19,8 @@ abstract class ApiService {
 
 @POST(ApiConstant.login)
 Future<LoginResponse> login(@Body()LoginRequestBody  loginRequestBody);
+
+@POST(ApiConstant.register)
+Future<SignupResponse> register(@Body()SignupRequestBody  signupRequestBody);
 
 }
