+++ lib/core/routing/app_router.dart
@@ -4,8 +4,9 @@ import 'package:doc_app/core/routing/roters.dart';
 import 'package:doc_app/feathures/home/ui/home.dart';
 import 'package:doc_app/feathures/login/logic/cubit/login_cubit.dart';
 
-import 'package:doc_app/feathures/login/logic/ui/login.dart';
+import 'package:doc_app/feathures/login/ui/login.dart';
 import 'package:doc_app/feathures/onbording/ui/onboarder.dart';
+import 'package:doc_app/feathures/sign_up/ui/sinup_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -26,6 +27,8 @@ class AppRouter {
         );
         case Routers.homeScreen:
         return MaterialPageRoute(builder: (_)=>const HomeScreen());
+        case Routers.signupScreen:
+        return MaterialPageRoute(builder: (_)=>const SinUpScreen());
 
       // Add more cases for other routes as needed
       default:
