+++ lib/feathures/login/logic/cubit/login_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:doc_app/feathures/login/logic/cubit/login_state.dart';
-import 'package:doc_app/feathures/login/logic/data/login_request_body.dart';
-import 'package:doc_app/feathures/login/logic/data/repo/login_repo.dart';
+import 'package:doc_app/feathures/login/data/login_request_body.dart';
+import 'package:doc_app/feathures/login/data/repo/login_repo.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -13,10 +13,10 @@ class LoginCubit extends Cubit<LoginState> {
   TextEditingController emailController = TextEditingController();
   TextEditingController passwordControl = TextEditingController();
 
-  void login(LoginRequestBody loginRequestBody) async {
+  void login() async {
     emit(const LoginState.loading());
 
-    final response = await _loginRepo.login(loginRequestBody);
+    final response = await _loginRepo.login(LoginRequestBody(email: emailController.text,password: passwordControl.text));
 
     response.when(success: (loginResponce) {
       emit(LoginState.success(loginResponce));
