+++ lib/feathures/login/ui/login.dart
@@ -2,12 +2,12 @@ import 'package:doc_app/core/helpers/spacing.dart';
 import 'package:doc_app/core/theming/style/styles.dart';
 import 'package:doc_app/core/widgets/text_button.dart';
 import 'package:doc_app/feathures/login/logic/cubit/login_cubit.dart';
-import 'package:doc_app/feathures/login/logic/data/login_request_body.dart';
-import 'package:doc_app/feathures/login/logic/ui/widgets/already_have_account.dart';
-import 'package:doc_app/feathures/login/logic/ui/widgets/condition_text.dart';
-import 'package:doc_app/feathures/login/logic/ui/widgets/email_and_password.dart';
-import 'package:doc_app/feathures/login/logic/ui/widgets/login_bloc_listener.dart';
-import 'package:doc_app/feathures/login/logic/ui/widgets/welcom_text.dart';
+import 'package:doc_app/feathures/login/data/login_request_body.dart';
+import 'package:doc_app/feathures/login/ui/widgets/already_have_account.dart';
+import 'package:doc_app/feathures/login/ui/widgets/condition_text.dart';
+import 'package:doc_app/feathures/login/ui/widgets/email_and_password.dart';
+import 'package:doc_app/feathures/login/ui/widgets/login_bloc_listener.dart';
+import 'package:doc_app/feathures/login/ui/widgets/welcom_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -58,9 +58,7 @@ class _LoginScreenState extends State<LoginScreen> {
 
   valiadateThenDoLogin() {
     if (context.read<LoginCubit>().formKey.currentState!.validate()) {
-      context.read<LoginCubit>().login(LoginRequestBody(
-          email: context.read<LoginCubit>().emailController.text,
-          password: context.read<LoginCubit>().passwordControl.text));
+      context.read<LoginCubit>().login();
     }
   }
 }
