+++ lib/feathures/login/ui/widgets/already_have_account.dart
@@ -1,0 +1,32 @@
+import 'package:doc_app/core/helpers/extensions.dart';
+import 'package:doc_app/core/routing/roters.dart';
+import 'package:doc_app/core/theming/style/styles.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+class AlredyHaveAccount extends StatelessWidget {
+  const AlredyHaveAccount({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return RichText(
+      textAlign: TextAlign.center,
+      text: TextSpan(
+        children: [
+          TextSpan(
+            text: 'Don\'t have an account yet? ',
+            //By logging, you agree to our    PrivacyPolicy.
+            style: Styles.font13Gryregular,
+          ),
+          TextSpan(
+              text: ' Sign Up ',
+              style: Styles.font13Buleregular,
+              recognizer: TapGestureRecognizer()
+                ..onTap = () {
+                  context.pushReplacementNamed(Routers.signupScreen);
+                }),
+        ],
+      ),
+    );
+  }
+}
